+++ lib/shark_on_lambda.rb
@@ -1,50 +1,47 @@
 # frozen_string_literal: true
 
+require 'pry' if Gem.loaded_specs.key?('pry')
+
 require 'forwardable'
 require 'ostruct'
 require 'pathname'
 require 'singleton'
 
+require 'action_controller'
+require 'action_dispatch'
+require 'action_view'
 require 'active_support/all'
 require 'jsonapi/deserializable'
 require 'jsonapi/serializable'
 require 'rack/utils'
 require 'yaml'
+require 'zeitwerk'
 
-require 'shark_on_lambda/version'
+# Without this, Zeitwerk hiccups in certain cases...
+module SharkOnLambda; end
 
-require 'shark_on_lambda/concerns/filter_actions'
-require 'shark_on_lambda/concerns/resettable_singleton'
-require 'shark_on_lambda/concerns/yaml_config_loader'
-
-require 'shark_on_lambda/configuration'
-require 'shark_on_lambda/secrets'
-
-require 'shark_on_lambda/inferrers/name_inferrer'
-require 'shark_on_lambda/inferrers/serializer_inferrer'
-
-require 'shark_on_lambda/api_gateway/serializers/base_error_serializer'
-require 'shark_on_lambda/api_gateway/concerns/http_response_validation'
-require 'shark_on_lambda/api_gateway/base_controller'
-require 'shark_on_lambda/api_gateway/base_handler'
-require 'shark_on_lambda/api_gateway/errors'
-require 'shark_on_lambda/api_gateway/headers'
-require 'shark_on_lambda/api_gateway/jsonapi_controller'
-require 'shark_on_lambda/api_gateway/jsonapi_parameters'
-require 'shark_on_lambda/api_gateway/jsonapi_renderer'
-require 'shark_on_lambda/api_gateway/parameters'
-require 'shark_on_lambda/api_gateway/query'
-require 'shark_on_lambda/api_gateway/request'
-require 'shark_on_lambda/api_gateway/response'
+Zeitwerk::Loader.for_gem.tap do |loader|
+  loader.ignore(File.expand_path('shark-on-lambda.rb', __dir__))
+  loader.inflector.inflect(
+    'rspec' => 'RSpec',
+    'version' => 'VERSION'
+  )
+  loader.setup
+  loader.eager_load
+end
 
 # Top-level module for this gem.
 module SharkOnLambda
   class << self
-    extend ::Forwardable
+    extend Forwardable
 
     attr_writer :logger
 
     def_instance_delegators :config, :root, :stage
+
+    def application
+      @application ||= Application.new
+    end
 
     def config
       Configuration.instance
@@ -72,7 +69,7 @@ module SharkOnLambda
     end
 
     def logger
-      @logger ||= ::Logger.new(STDOUT)
+      @logger ||= Logger.new(STDOUT)
     end
 
     def reset_configuration