+++ lib/shark_on_lambda/response.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 module SharkOnLambda
-  VERSION = '1.0.0.rc1'
+  class Response < ActionDispatch::Response
+  end
 end